+++ time_warp/src/shared/delta.rs
@@ -1,6 +1,6 @@
 #[derive(Clone)]
 pub struct Delta {
-  z: f32,
+  z: f64,
 }
 
 impl Delta {
@@ -8,7 +8,7 @@ impl Delta {
     Self { z: 0. }
   }
 
-  pub fn process(&mut self, input: f32) -> f32 {
+  pub fn process(&mut self, input: f64) -> f64 {
     let output = input - self.z;
     self.z = input;
     output
